requirements.txt
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
Pillow>=10.0.0
PyQt6>=6.5.0 